add custom viewcontroller for scroll
